move core into top-level package
