+++ src/magql/search.py
