+++ src/magql/sqlalchemy/manager.py
@@ -7,12 +7,12 @@ import sqlalchemy as sa
 import sqlalchemy.orm as sa_orm
 from sqlalchemy.sql.type_api import TypeEngine
 
-from ..core import nodes
-from ..core import scalars
-from ..core.filters import filter_item
-from ..core.schema import Schema
-from ..core.search import Search
-from ..core.search import SearchProvider
+from .. import nodes
+from .. import scalars
+from ..filters import filter_item
+from ..schema import Schema
+from ..search import Search
+from ..search import SearchProvider
 from .check_delete import CheckDelete
 from .pagination import PerPageValidator
 from .pagination import validate_page
